fix NPE when trying to wipe cached files from the disk

+++ src/main/java/com/github/droidfu/cachefu/AbstractCache.java
@@ -315,6 +315,9 @@ public abstract class AbstractCache<KeyT, ValT> implements Map<KeyT, ValT> {
 
         if (isDiskCacheEnabled) {
             File[] cachedFiles = new File(diskCacheDirectory).listFiles();
+            if (cachedFiles == null) {
+                return;
+            }
             for (File f : cachedFiles) {
                 f.delete();
             }
